Bluetooth: Mesh: Sensor Server and Client sample
Adds an application sample for Bluetooth Mesh Sensor Server and
an application sample for Bluetooth Mesh Sensor Client. These two
samples are made to work together, where the Server provides sensor
data to the Client. The Bluetooth Mesh Sensor framework is used.
Four Bluetooth Mesh sensor types are used:

- bt_mesh_sensor_present_dev_op_temp
- bt_mesh_sensor_rel_runtime_in_a_dev_op_temp_range
- bt_mesh_sensor_presence_detected
- bt_mesh_sensor_time_since_presence_detected

On the Server the on-chip TEMP_NRF5 temperature sensor is used
in addidion to Button 1 on the DK simulating presence.

Signed-off-by: Erik Robstad <Erik.Robstad@nordicsemi.no>